@@ -1,13 +1,12 @@
 import axios from "axios";
-import {api, apiUrl} from "../assets/js/config.js";
-import {userInfo} from "../Store/userInfo.js";
-import {projectReviewsStore} from "../Store/projectReviews.js";
-import re from "floating-vue";
-import {projectsStore} from "../Store/projectsStore.js";
+import {api} from "@/API/apiurl.js";
+import {userStore} from "@/Stores/userStore.js";
+import {addNotice} from "@/js/notifications.js";
+import {popup} from "@/js/controllers/popupController.js";
 
 export function getReviews(options) {
     const headers = {
-        'Authorization': `Bearer ${userInfo.token}`
+        'Authorization': `Bearer ${userStore.token}`
     };
     let link = '?'
     if (options) {
@@ -16,28 +15,36 @@ export function getReviews(options) {
         }
     }
     link = link.slice(0, -1)
-    return axios.get(`${apiUrl}reviews${link}`, {headers})
+    return axios.get(`${api.url}reviews${link}`, {headers})
 
 }
 
 
 export async function addReview(review) {
     const headers = {
-        'Authorization': `Bearer ${userInfo.token}`
+        'Authorization': `Bearer ${userStore.token}`
     };
 
-
-
-    return await axios.post(`${apiUrl}reviews`, review, {headers})
+    return await axios.post(`${api.url}reviews`, review, {headers})
         .then(result => {
-
-        }).catch(error => error)
+            addNotice({name: 'Отзыв отправлен на модерацию', type: 'success'})
+            popup.show = false
+        }).catch(error => {
+                let message;
+                if (error.response.data.message === "This user already rated this project.") {
+                    message = 'Вы уже оставляли отзыв'
+                } else {
+                    message = 'Ошибка :('
+                }
+                addNotice({name: message, type: 'danger'})
+            }
+        )
 }
 
 
 export function showNotReviewed(options) {
     const headers = {
-        'Authorization': `Bearer ${userInfo.token}`
+        'Authorization': `Bearer ${userStore.token}`
     };
     let link = ''
     if (options) {
@@ -46,15 +53,15 @@ export function showNotReviewed(options) {
         }
     }
     link = link.slice(0, -1)
-    return axios.get(`${apiUrl}reviews?showNotReviewed=true&${link}`, {headers})
+    return axios.get(`${api.url}reviews?showNotReviewed=true&${link}`, {headers})
 }
 
 export async function editReview(review) {
     const headers = {
-        'Authorization': `Bearer ${userInfo.token}`
+        'Authorization': `Bearer ${userStore.token}`
     };
 
-    return await axios.put(`${apiUrl}reviews/${review.id}`, review, {headers})
+    return await axios.put(`${api.url}reviews/${review.id}`, review, {headers})
         .then(result => {
             const review = projectReviewsStore.reviews.find(item => item.id === result.data.updatedReview.id)
             review.comment = result.data.updatedReview.comment
@@ -64,32 +71,26 @@ export async function editReview(review) {
 
 export async function deleteReview(review) {
     const headers = {
-        'Authorization': `Bearer ${userInfo.token}`
+        'Authorization': `Bearer ${userStore.token}`
     };
-    projectReviewsStore.reviews.splice(projectReviewsStore.reviews.findIndex(item => item.id === review.id), 1)
-    return await axios.delete(`${apiUrl}reviews/${review.id}`, {headers})
-        .then(result => result).catch(error => error)
+    return axios.delete(`${api.url}reviews/${review.id}`, {headers})
 }
 
 export async function approveReview(review) {
     const headers = {
-        'Authorization': `Bearer ${userInfo.token}`
+        'Authorization': `Bearer ${userStore.token}`
     };
     review.isReviewed = true
 
-    if (projectsStore.projects.find(item => item.id === review.ProjectId)) {
-        projectsStore.projects.find(item => item.id === review.ProjectId).reviewsCount =  projectsStore.projects.find(item => item.id === review.ProjectId).reviewsCount + 1
-    }
-    return await axios.put(`${api.url}reviews/${review.id}`, review, {headers})
-        .then(result => result).catch(error => error.response)
+    return axios.put(`${api.url}reviews/${review.id}`, review, {headers})
 }
 
 export async function disapproveReview(review) {
     const headers = {
-        'Authorization': `Bearer ${userInfo.token}`
+        'Authorization': `Bearer ${userStore.token}`
     };
     review.isReviewed = false
 
-    return await axios.put(`${apiUrl}reviews/${review.id}`, review, {headers})
+    return await axios.put(`${api.url}reviews/${review.id}`, review, {headers})
         .then(result => result).catch(error => error)
 }
